+++ res/html/problem.js
@@ -4,6 +4,7 @@ window.addEventListener('DOMContentLoaded', () => {
     window.registerTab('Solution');
     window.registerTab('Related');
     window.registerTab('Files');
+    window.registerTab('Record');
 
     window.setMessageHandler((message) => {
         const data = message.data;
@@ -268,6 +269,44 @@ window.addEventListener('DOMContentLoaded', () => {
                 }
                 window.enableTab('Solution', solutionHTML);
                 window.setTabCount('Solution', data.psdocs.length);
+                break;
+            }
+            case 'record': {
+                if (message.error) {
+                    window.enableTab('Record', `<div class="center error">
+                        <vscode-icon name="error" size="18"></vscode-icon>
+                        <div>${message.error}</div>
+                    </div>`);
+                    return;
+                }
+                var recordHTML = `
+                <vscode-table zebra bordered-columns responsive resizable breakpoint="400">
+                    <vscode-table-header slot="header">
+                        <vscode-table-header-cell>Status</vscode-table-header-cell>
+                        <vscode-table-header-cell>Submit By</vscode-table-header-cell>
+                        <vscode-table-header-cell>Time</vscode-table-header-cell>
+                        <vscode-table-header-cell>Memory</vscode-table-header-cell>
+                        <vscode-table-header-cell>Language</vscode-table-header-cell>
+                        <vscode-table-header-cell>Submit At</vscode-table-header-cell>
+                    </vscode-table-header>
+                    <vscode-table-body slot="body">`;
+                for (const rdoc of data.rdocs) {
+                    recordHTML += `<vscode-table-row>
+                        <vscode-table-cell style="cursor: pointer" onclick="vscode.postMessage({command:'openT',data:['${rdoc._id}']})">
+                            <span class="icon record-status--icon ${window.statusIcon[rdoc.status]}"></span>
+                            <span style="color: ${window.scoreColor[Math.floor(rdoc.score / 100 * 10)]}">${rdoc.score}</span>
+                            <span class="record-status--text ${window.statusIcon[rdoc.status]}">${window.statusName[rdoc.status]}</span>
+                        </vscode-table-cell>
+                        <vscode-table-cell>${data.udict[rdoc.uid].uname}</vscode-table-cell>
+                        <vscode-table-cell>${window.toTime(rdoc.time)}</vscode-table-cell>
+                        <vscode-table-cell>${window.toMemory(rdoc.memory * 1024)}</vscode-table-cell>
+                        <vscode-table-cell>${window.languageDisplayName[rdoc.lang]}</vscode-table-cell>
+                        <vscode-table-cell>${window.toRelativeTime(new Date(rdoc.judgeAt).getTime())}</vscode-table-cell>
+                    </vscode-table-row>`;
+                }
+                recordHTML += `</vscode-table-body></vscode-table>`;
+                window.enableTab('Record', recordHTML);
+                window.setTabCount('Record', data.rdocs.length);
                 break;
             }
             default: